% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.dt}
\alias{test.dt}
\title{Testing Dataset}
\format{
the testing data
\describe{
\item{Yij}{survial time}
\item{Xij}{covaiates}
\item{delta}{event}
\item{Mi}{Location}
\item{d2}{distance matrix}
}
}
\usage{
test.dt
}
\description{
A dataset containing survival analysis data for model testing.
}
\keyword{datasets}
