% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train.dt}
\alias{train.dt}
\title{Training Dataset}
\format{
A list object containing the following components:
\describe{
\item{Yij}{Survival time for each subject}
\item{Xij}{Covariate matrix}
\item{delta}{Event indicator (1 = event occurred, 0 = censored)}
\item{Mi}{Spatial location of each subject}
\item{d2}{Distance matrix between locations}
}
}
\usage{
train.dt
}
\description{
A dataset containing survival analysis data for model training.
}
\keyword{datasets}
